+++ CHANGELOG.md
@@ -52,6 +52,7 @@ v5.5.0
 - Add three subsonic inflow boundary conditions: EnforceEntropyPresure, EnforceEntropyVelocity, and EnforceEntropyTotalEnthalpy (!1398)
 - Update implicit solver for implicit PFASST (!1659)
 - Some tidy-up in the compressible flow solver (!1698)
+- Fix ESDIRK scheme for compressible flow solver (!1690)
 
 **ShallowWaterSolver**
 - Reduce memory footprint of the Peregrine solver(!1680)